bacery/main/api/serializers.py
from rest_framework import serializers
from main.models import Subscriber


class SubscriberSerializer(serializers.ModelSerializer):
    class Meta:
        model = Subscriber
        fields = ('email',)




